+++ src/components/posts/UsersPosts.tsx
@@ -1,30 +1,16 @@
 import { Text } from "@chakra-ui/react";
-import styled from "@emotion/styled";
-import {
-  useInfiniteQuery,
-  useQuery,
-  useQueryClient,
-} from "@tanstack/react-query";
-import React, { useEffect, useRef, useState } from "react";
-import { QueryPostDto } from "../../api/dtos/post.dto";
+import { useInfiniteQuery } from "@tanstack/react-query";
+import React, { useEffect, useRef } from "react";
 import postApi from "../../api/postApi";
 import BlueSpinner from "../../common/components/spinners/BlueSpinner";
+
 import Post from "./Post";
-import PostV2 from "./PostV2";
 
 interface UsersPostsProps {
   uids: number[];
 }
 const UsersPosts = ({ uids }: UsersPostsProps) => {
-  const {
-    data,
-    error,
-    fetchNextPage,
-    hasNextPage,
-    isFetching,
-    isFetchingNextPage,
-    status,
-  } = useInfiniteQuery({
+  const { data, fetchNextPage, isFetching } = useInfiniteQuery({
     queryKey: ["posts", "followings"],
     queryFn: async ({ pageParam = 0 }) => {
       const posts = await postApi.getUsersPosts({
@@ -63,7 +49,7 @@ const UsersPosts = ({ uids }: UsersPostsProps) => {
       {data?.pages.map((page, i) => (
         <React.Fragment key={i}>
           {page.map((post) => (
-            <PostV2 pid={post.id} />
+            <Post pid={post.id} />
           ))}
         </React.Fragment>
       ))}
