+++ src/pages/post/SinglePost.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import PostV2 from "../../components/posts/PostV2";
+import PostV2 from "../../components/posts/Post";
 
 const SinglePost = () => {
   const slug = useParams().slug;
